team-research/requirements.txt
pyautogen
openai>=1